audio_processor/requirements.txt
pyaudio
SpeechRecognition
psycopg2-binary
numpy
torch
sentence-transformers
